refactoring old code to openaiwrapper to make room for openai
