+++ backend/lucidify-api/middleware/logging.go
