+++ backend/lucidify-api/openaiwrapper/chatmodel/chat_completion_model.go
