+++ backend/lucidify-api/openaiwrapper/chatmodel/chat_completion_model_integration_test.go
