+++ backend/lucidify-api/openaiwrapper/chatthread/chat_thread_unit_test.go
@@ -1,7 +1,7 @@
 package chatthread
 
 import (
-	"openaiwrapper-integrations/openaiwrapper/chatmodel"
+	"lucidify-api/openaiwrapper/chatmodel"
 	"testing"
 )
