+++ backend/lucidify-api/openaiwrapper/sales_utils.go
