--- backend/openai-cookbook-q-and-a/.gitignore
@@ -1,2 +1,0 @@
-__pycache__
-env
