--- backend/openai-cookbook-q-and-a/.vimspector.json
@@ -1,28 +1,0 @@
-{
-  "configurations": {
-    "Flask (debug)": {
-      "adapter": "debugpy",
-      "configuration": {
-        "name": "Flask",
-        "type": "python",
-        "request": "launch",
-        "program": "${workspaceRoot}/app.py",
-        "cwd": "${workspaceRoot}",
-        "env": {
-          "FLASK_ENV": "development",
-          "FLASK_APP": "app.py"
-        },
-        "args": [
-          "run",
-          "--no-debugger",
-          "--no-reload",
-          "--host=0.0.0.0",
-          "--port=5000"
-        ],
-        "justMyCode": false,
-        "stopOnEntry": false
-      }
-    }
-  }
-}
-
