--- backend/openai-cookbook-q-and-a/Dockerfile.dev
@@ -1,22 +1,0 @@
-# Use an official Python runtime as the parent image
-FROM python:3.9-slim
-
-# Set the working directory in the container to /app
-WORKDIR /app
-
-# Copy the current directory contents into the container at /app
-COPY requirements.txt ./
-
-# Install any needed packages specified in requirements.txt
-RUN pip install --no-cache-dir -r requirements.txt
-
-# Make port 8080 available to the world outside this container
-EXPOSE 8080
-
-# Set environment variables
-ENV FLASK_APP=app.py
-ENV FLASK_RUN_HOST=0.0.0.0
-ENV FLASK_ENV=development
-
-# Run app.py when the container launches
-CMD ["flask", "run", "--port=8080"]
